Enhetstester for DigipostActionFactory
